[refactor] 패키지 구조 수정 및 로그 확인 위해 주석 처리 풀기
