+++ src/main/java/store/greeting/config/WebSocketConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-import store.greeting.cart.ChatSocketHandler;
+import store.greeting.chat.ChatSocketHandler;
 
 @Configuration
 @RequiredArgsConstructor
